Fix cross-platform path string conversion
